# api/openapi-fedramp-dashboard.yaml
openapi: 3.0.3
info:
  title: FedRAMP Security Dashboard API
  description: |
    REST API for the FedRAMP Security Dashboard - Phase 2 of 5
    
    Provides real-time compliance status, historical trends, and control validation data
    with role-based access control (RBAC) for security teams.
    
    **Data Sources:**
    - Azure Monitor (real-time metrics)
    - Cosmos DB (90-day hot storage)
    - Log Analytics (KQL queries)
    
    **Authentication:** Azure AD / Microsoft Entra ID
    
    **RBAC Roles:**
    - Security Admin (full access)
    - Security Engineer (read/write validation data)
    - SRE (operational dashboards)
    - Ops Viewer (read-only dashboards)
    - Auditor (compliance reports only)
  version: 2.0.0
  contact:
    name: Platform Security Team
    email: platform-security@contoso.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://fedramp-dashboard-api-prod.azurewebsites.net/api/v1
    description: Production
  - url: https://fedramp-dashboard-api-stg.azurewebsites.net/api/v1
    description: Staging
  - url: https://localhost:5001/api/v1
    description: Local development
security:
  - AzureAD: []
...
paths:
...
  /reports/compliance-export:
    get:
      summary: Export compliance report
      description: |
        Generates a compliance report in CSV or JSON format.
        Suitable for audit documentation and external reporting.
        
        **Permission:** `Reports.Export` (Security Admin, Security Engineer, Auditor)
      operationId: exportComplianceReport
      tags:
        - Reports
      parameters:
        - name: format
          in: query
          description: Export format
          required: false
          schema:
            type: string
            enum: [json, csv]
            default: json
        - name: environment
          in: query
          description: Environment filter
          required: false
          schema:
            type: string
            enum: [DEV, STG, PROD, ALL]
            default: ALL
        - name: startDate
          in: query
          description: Report start date
          required: true
          schema:
            type: string
            format: date-time
        - name: endDate
          in: query
          description: Report end date
          required: true
          schema:
            type: string
            format: date-time
        - name: includeDetails
          in: query
          description: Include detailed test results
          required: false
          schema:
            type: boolean
            default: false
      responses:
        '200':
          description: Compliance report generated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ComplianceReport'
            text/csv:
              schema:
                type: string
                format: binary
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '403':
          $ref: '#/components/responses/Forbidden'
        '500':
          $ref: '#/components/responses/InternalServerError'
      security:
        - AzureAD:
            - Reports.Export
components:
  securitySchemes:
    AzureAD:
      type: oauth2
      description: Azure Active Directory / Microsoft Entra ID authentication
      flows:
        implicit:
          authorizationUrl: https://login.microsoftonline.com/{tenant}/oauth2/v2.0/authorize
          scopes:
            Dashboard.Read: Read access to compliance dashboards
            Controls.Read: Read access to control validation data
            Analytics.Read: Read access to analytics and drift detection
            Reports.Export: Export compliance reports
  schemas:
...
    ComplianceReport:
      type: object
      properties:
        report_id:
          type: string
        generated_at:
          type: string
          format: date-time
        report_period:
          type: object
          properties:
            start_date:
              type: string
              format: date-time
            end_date:
              type: string
              format: date-time
        environments:
          type: array
          items:
            type: string
        summary:
          type: object
          properties:
            total_tests:
              type: integer
            passed_tests:
              type: integer
            failed_tests:
              type: integer
            overall_compliance_rate:
              type: number
              format: float
        control_results:
          type: array
          items:
            type: object
            properties:
              control_id:
                type: string
              control_name:
                type: string
              pass_count:
                type: integer
              fail_count:
                type: integer
              compliance_rate:
                type: number
                format: float
...
    Error:
      type: object
      properties:
        error_code:
          type: string
        message:
          type: string
        timestamp:
          type: string
          format: date-time
        trace_id:
          type: string
  responses:
    BadRequest:
      description: Invalid request parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Unauthorized:
      description: Authentication required
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
    Forbidden:
      description: Insufficient permissions for this operation
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
...
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'